+++ tests/integration/test_components.py
@@ -36,7 +36,10 @@ gi.require_version("GLib", "2.0")
 gi.require_version("GtkSource", "4")
 from gi.repository import Gtk, GLib, GtkSource, Gdk
 
-from tests.test import quick_cleanup, spy, logger
+from tests.cleanup import quick_cleanup
+from tests.stuff import spy
+from tests.logger import logger
+
 from inputremapper.input_event import InputEvent
 from inputremapper.gui.utils import CTX_ERROR, CTX_WARNING, gtk_iteration
 from inputremapper.gui.messages.message_broker import (
