+++ tests/test.py
@@ -77,25 +77,18 @@ if __name__ == "__main__":
 
     tests.test.main()
 
-import shutil
 import time
-import copy
 import unittest
 import subprocess
-import asyncio
 import psutil
-from pickle import UnpicklingError
-from unittest.mock import patch
 
 os.environ["UNITTEST"] = "1"
 
 from tests.logger import logger
 from tests.constants import EVENT_READ_TIMEOUT
-from tests.tmp import tmp
 from tests.fixtures import fixtures
 from tests.pipes import (
     setup_pipe,
-    pending_events,
     uinput_write_history,
     uinput_write_history_pipe,
 )
@@ -107,9 +100,9 @@ from tests.patches import (
     patch_regrab_timeout,
     patch_is_running,
     patch_evdev,
-    uinputs,
     InputEvent,
 )
+from tests.cleanup import cleanup
 
 
 def is_service_running():
@@ -184,126 +177,8 @@ patch_is_running()
 # patch_warnings()
 
 
-environ_copy = copy.deepcopy(os.environ)
-
-
-def quick_cleanup(log=True):
-    """Reset the applications state."""
-    from inputremapper.gui.utils import debounce_manager
-    from inputremapper.configs.global_config import global_config
-    from inputremapper.injection.macros.macro import macro_variables
-    from inputremapper.configs.system_mapping import system_mapping
-    from inputremapper.configs.paths import get_config_path
-    from inputremapper.injection.global_uinputs import global_uinputs
-
-    if log:
-        print("Quick cleanup...")
-
-    debounce_manager.stop_all()
-
-    for device in list(pending_events.keys()):
-        try:
-            while pending_events[device][1].poll():
-                pending_events[device][1].recv()
-        except (UnpicklingError, EOFError):
-            pass
-
-        # setup new pipes for the next test
-        pending_events[device][1].close()
-        pending_events[device][0].close()
-        del pending_events[device]
-        setup_pipe(device)
-
-    try:
-        if asyncio.get_event_loop().is_running():
-            for task in asyncio.all_tasks():
-                task.cancel()
-    except RuntimeError:
-        # happens when the event loop disappears for magical reasons
-        # create a fresh event loop
-        asyncio.set_event_loop(asyncio.new_event_loop())
-
-    if macro_variables.process is not None and not macro_variables.process.is_alive():
-        # nothing should stop the process during runtime, if it has been started by
-        # the injector once
-        raise AssertionError("the SharedDict manager is not running anymore")
-
-    if macro_variables.process is not None:
-        macro_variables._stop()
-
-    join_children()
-
-    macro_variables.start()
-
-    if os.path.exists(tmp):
-        shutil.rmtree(tmp)
-
-    global_config.path = os.path.join(get_config_path(), "config.json")
-    global_config.clear_config()
-    global_config._save_config()
-
-    system_mapping.populate()
-
-    clear_write_history()
-
-    for name in list(uinputs.keys()):
-        del uinputs[name]
-
-    # for device in list(active_macros.keys()):
-    #    del active_macros[device]
-    # for device in list(unreleased.keys()):
-    #    del unreleased[device]
-    fixtures.reset()
-    os.environ.update(environ_copy)
-    for device in list(os.environ.keys()):
-        if device not in environ_copy:
-            del os.environ[device]
-
-    for _, pipe in pending_events.values():
-        assert not pipe.poll()
-
-    assert macro_variables.is_alive(1)
-    for uinput in global_uinputs.devices.values():
-        uinput.write_count = 0
-        uinput.write_history = []
-
-    global_uinputs.is_service = True
-
-    if log:
-        print("Quick cleanup done")
-
-
-def cleanup():
-    """Reset the applications state.
-
-    Using this is slower, usually quick_cleanup() is sufficient.
-    """
-    from inputremapper.groups import groups
-    from inputremapper.injection.global_uinputs import global_uinputs
-
-    print("Cleanup...")
-
-    os.system("pkill -f input-remapper-service")
-    os.system("pkill -f input-remapper-control")
-    time.sleep(0.05)
-
-    quick_cleanup(log=False)
-    groups.refresh()
-    with patch.object(sys, "argv", ["input-remapper-service"]):
-        global_uinputs.prepare_all()
-
-    print("Cleanup done")
-
-
-def spy(obj, name):
-    """Convenient wrapper for patch.object(..., ..., wraps=...)."""
-    return patch.object(obj, name, wraps=obj.__getattribute__(name))
-
-
-cleanup()
-
-
 def main():
+    cleanup()
     # https://docs.python.org/3/library/argparse.html
     parser = argparse.ArgumentParser(description=__doc__)
     # repeated argument 0 or more times with modules
