+++ tests/unit/test_config.py
@@ -19,7 +19,8 @@
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from tests.test import quick_cleanup, tmp
+from tests.cleanup import quick_cleanup
+from tests.tmp import tmp
 
 import os
 import unittest
