+++ tests/unit/test_groups.py
@@ -19,7 +19,7 @@
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from tests.test import quick_cleanup
+from tests.cleanup import quick_cleanup
 from tests.fixtures import fixtures, keyboard_keys
 
 import os
