+++ tests/unit/test_reader.py
@@ -51,9 +51,9 @@ from inputremapper.gui.messages.message_types import MessageType
 from inputremapper.gui.reader_client import ReaderClient
 from inputremapper.gui.reader_service import ReaderService
 from tests.test import (
-    new_event,
-    quick_cleanup,
+    new_event
 )
+from tests.cleanup import quick_cleanup
 from tests.constants import (
     EVENT_READ_TIMEOUT,
     START_READING_DELAY,
