+++ tests/unit/test_test.py
@@ -19,7 +19,8 @@
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
 from inputremapper.gui.messages.message_broker import MessageBroker
-from tests.test import quick_cleanup, cleanup, new_event
+from tests.test import new_event
+from tests.cleanup import cleanup, quick_cleanup
 from tests.constants import EVENT_READ_TIMEOUT, START_READING_DELAY
 from tests.logger import logger
 from tests.fixtures import fixtures
