+++ tests/unit/test_user.py
@@ -19,7 +19,7 @@
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from tests.test import quick_cleanup
+from tests.cleanup import quick_cleanup
 
 import os
 import unittest
