+++ tests/xmodmap.py
@@ -1,3 +1,23 @@
+#!/usr/bin/python3
+# -*- coding: utf-8 -*-
+# input-remapper - GUI for device specific keyboard mappings
+# Copyright (C) 2022 sezanzeb <proxima@sezanzeb.de>
+#
+# This file is part of input-remapper.
+#
+# input-remapper is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# input-remapper is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
+
 xmodmap = (
     b"keycode   8 =\nkeycode   9 = Escape NoSymbol Escape\nkeycode  10 = 1 exclam 1 exclam onesuperior exclamdown ones"
     b"uperior\nkeycode  11 = 2 quotedbl 2 quotedbl twosuperior oneeighth twosuperior\nkeycode  12 = 3 section 3 sectio"
